First commit
A port of an existing project. Just a few days ago made possible after the maintainer of Adafruit tinyuf2 added the ESP32-S3-Box, on my request, to the list of boards. After this I was able to build a board port of tinyuf2 bootloader for the ESP32-S3-Box. Then I was able to copy CircuitPython V8.0.0-beta.2 to this 'esp-box' and then it was possible to port this project to the esp-box.